requirements.txt
aiofiles==23.2.1
asgiref==3.7.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
grpcio==1.62.0
grpcio-tools==1.62.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pika==1.3.2
pip==23.3.1
protobuf==4.25.3
requests==2.31.0
setuptools==68.2.2
urllib3==2.2.1
Werkzeug==3.0.1